+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [Core] Restrict the maximum number of unhealthy nodes in MachineHealthCheck
 * [Core] Add PDB and PriorityClass to capx components
 * [Core] Bump cluster api to v1.5.3
+* [AWS][EKS] Support aws load balancer controller manager
 * [Core] Enable scale from zero
 * [Core] Add core dns PDB if required
 * [Core] Add keos 1.1.x support